Suggestion System Complete

+++ application/views/showall.php
@@ -21,13 +21,15 @@
                         <div class="row">
                             <div class="col-sm-10">
                                 <p class="form-control-static"><?php echo $item->title ?></p>
-                                <input type="edit" value="<?php echo $item->title ?>" class="hide" />
+                                <input type="edit" value="<?php echo $item->title ?>" class="hide" data-type="title" data-id="<?php echo $item->id ?>"/>
                             </div>
-                            <div class="col-sm-2">
-                                <a href="#">
-                                    <span class="glyphicon glyphicon-pencil hide" title="Edit"></span>
-                                </a>
-                            </div>
+                            <?php if ($authenticated): ?>
+                                <div class="col-sm-2">
+                                    <a href="#">
+                                        <span class="glyphicon glyphicon-pencil hide" title="Edit"></span>
+                                    </a>
+                                </div>
+                            <?php endif; ?>
                         </div>
                     </div>
                 </td>
@@ -36,13 +38,15 @@
                         <div class="row">
                             <div class="col-sm-10">
                                 <p class="form-control-static"><?php echo $item->artist ?></p>
-                                <input type="edit" value="<?php echo $item->artist ?>" class="hide" />
+                                <input type="edit" value="<?php echo $item->artist ?>" class="hide" data-type="artist" data-id="<?php echo $item->id ?>" />
                             </div>
-                            <div class="col-sm-2">
-                                <a href="#">
-                                    <span class="glyphicon glyphicon-pencil hide"></span>
-                                </a>
-                            </div>
+                            <?php if ($authenticated): ?>
+                                <div class="col-sm-2">
+                                    <a href="#">
+                                        <span class="glyphicon glyphicon-pencil hide" title="Edit"></span>
+                                    </a>
+                                </div>
+                            <?php endif; ?>
                         </div>
                     </div>
                 </td>
@@ -51,21 +55,25 @@
                         <div class="row">
                             <div class="col-sm-10">
                                 <p class="form-control-static"><?php echo $item->album ?></p>
-                                <input type="edit" value="<?php echo $item->album ?>" class="hide" />
+                                <input type="edit" value="<?php echo $item->album ?>" class="hide" data-type="album" data-id="<?php echo $item->id ?>"/>
                             </div>
-                            <div class="col-sm-2">
-                                <a href="#">
-                                    <span class="glyphicon glyphicon-pencil hide"></span>
-                                </a>
-                            </div>
+                            <?php if ($authenticated): ?>
+                                <div class="col-sm-2">
+                                    <a href="#">
+                                        <span class="glyphicon glyphicon-pencil hide" title="Edit"></span>
+                                    </a>
+                                </div>
+                            <?php endif; ?>
                         </div>
                     </div>
                 </td>
                 <td>
-                    <div class="btn-group">
-                        <a href="editinfo/<?php echo $item->id ?>" class="btn btn-info" >EDIT</a>
-                        <a href="deleteinfo/<?php echo $item->id ?>" class="btn btn-danger">DELETE</a>
-                    </div>
+                    <?php if ($authenticated): ?>
+                        <div class="btn-group">
+                            <a href="editinfo/<?php echo $item->id ?>" class="btn btn-info" >EDIT</a>
+                            <a href="deleteinfo/<?php echo $item->id ?>" class="btn btn-danger">DELETE</a>
+                        </div>
+                    <?php endif; ?>
                 </td>
             </tr>
         <?php endforeach; ?>
